+++ src/Field/FieldBase.hh
@@ -51,20 +51,20 @@ public:
   void unregisterNodeList();
 
   // Methods every field must provide.
-  virtual unsigned size() const = 0;
+  virtual size_t size() const = 0;
   virtual void Zero() = 0;
   virtual void setNodeList(const NodeList<Dimension>& nodeList) = 0;
-  virtual std::vector<char> packValues(const std::vector<int>& nodeIDs) const = 0;
-  virtual void unpackValues(const std::vector<int>& nodeIDs,
+  virtual std::vector<char> packValues(const std::vector<size_t>& nodeIDs) const = 0;
+  virtual void unpackValues(const std::vector<size_t>& nodeIDs,
                             const std::vector<char>& buffer) = 0;
-  virtual void copyElements(const std::vector<int>& fromIndices,
-                            const std::vector<int>& toIndices) = 0;
+  virtual void copyElements(const std::vector<size_t>& fromIndices,
+                            const std::vector<size_t>& toIndices) = 0;
   virtual bool fixedSizeDataType() const = 0;
-  virtual int numValsInDataType() const = 0;
-  virtual int sizeofDataType() const = 0;
-  virtual int computeCommBufferSize(const std::vector<int>& packIndices,
-                                    const int sendProc,
-                                    const int recvProc) const = 0;
+  virtual size_t numValsInDataType() const = 0;
+  virtual size_t sizeofDataType() const = 0;
+  virtual size_t computeCommBufferSize(const std::vector<size_t>& packIndices,
+                                       const int sendProc,
+                                       const int recvProc) const = 0;
 
 //   // Methods to support cacheing of coarse and refine neighbor values.
 //   void notifyNewCoarseNodes() const;
@@ -75,11 +75,11 @@ protected:
   void setFieldBaseNodeList(const NodeList<Dimension>& nodeListPtr);
 
   friend class NodeList<Dimension>;
-  virtual void resizeField(unsigned size) = 0;
-  virtual void resizeFieldInternal(unsigned size, unsigned oldFirstGhostNode) = 0;
-  virtual void resizeFieldGhost(unsigned size) = 0;
-  virtual void deleteElement(int nodeID) = 0;
-  virtual void deleteElements(const std::vector<int>& nodeIDs) = 0;
+  virtual void resizeField(size_t size) = 0;
+  virtual void resizeFieldInternal(size_t size, size_t oldFirstGhostNode) = 0;
+  virtual void resizeFieldGhost(size_t size) = 0;
+  virtual void deleteElement(size_t nodeID) = 0;
+  virtual void deleteElements(const std::vector<size_t>& nodeIDs) = 0;
 
   // Make the FieldListBase a friend, so that it can use the registration
   // methods.
